refactor: Mode 클래스 수정(이름 변경, 파일 위치 변경, 리포지토리와의 의존성 제거)
